+++ src/ids-menu/ids-separator.d.ts
@@ -1,0 +1,5 @@
+// Ids is a JavaScript project, but we define TypeScript declarations so we can
+// confirm our code is type safe, and to support TypeScript users.
+import { IdsElement } from '../ids-base/ids-element';
+
+export default class IdsSeparator extends IdsElement {}